Implement sync operation for develop dependencies
Sync operation which updates the develop mode dependencies to the
version specified in the lock file is implemented. If the specified
revision is not found locally tries to fetch it from the configured
remotes. If it is present on multiple branches tries to stay on the
current one and if cannot prefers local branches rather than remote
tracking ones. If found on more than one branch gives the user a choice
whether to switch.

Sync operation will also download non develop mode dependencies'
versions described in the lock file if they are not already present in
the Nimble cache.

If `-l, --list-only` option is given than only list develop mode
dependencies which working copies are out of sync without actually
syncing them and without downloading missing non develop mode
dependencies.

**Important implementation details:**

To be able to determine whether a working copy of develop mode
dependency needs to be synced, locked again or merged with or re-based
on some other branch a special sync file is kept is the VCS directory
(.git or .hg) of the current package. It keeps a record for every
develop mode dependency for its current working copy revision during the
last `lock`, `sync` or `develop` operation. The name of the file is
`<package_name>.nimble.sync`.

**Additional changes:**

 - Nimble package directories which are not under version control, with
   not clean working copies or with not pushed VCS revisions cannot be
   locked.

 - Implement updating of already existing lock file. The update is being
   executed by the following algorithm:

  * Add newly added dependencies.

  * Remove missing dependencies.

  * Update develop mode dependencies to the newly locked version.

  * Leave installed dependencies to the previously locked version.

 - On `check` command the develop mode dependencies are being validated
   against the lock file. The following reasons for validation failure
   are possible.

  * The package directory is not under version control.

  * The package working copy directory is not in clean state.

  * Current VCS revision is no pushed on any remote.

  * The working copy needs sync. This happens when the VCS revision
    written in the sync file is equal to the current working copy
    revision, but the revision in the lock file is different from them.

  * The working copy needs lock. This happens when the lock file VCS
    revision is equal to the VCS revision from the sync file but the
    current working copy VCS revision is different.

  * The working copy needs merge or re-base. This happens when lock
    file, sync file and current working copy VCS revisions are all
    different.

 - Implemented caching of loaded develop file to avoid multiple loads
   when information for them is queried from different places in the
   code.

Related to nim-lnag/nimble#127