Enhance `develop --with-dependencies`
Make `develop --with-dependencies` to skip already existing directories
when cloning the repositories. This is useful when a second
`develop --with-dependencies` is executed for another package and some
of the dependencies are the same as in the first run. In that case, we
don't want the entire command to fail because some package directories
already exist.

Related to nim-lang/nimble#127